better tensorrt code
