+++ convert_onnx.py
@@ -1,26 +1,22 @@
 import argparse
 import logging
-import multiprocessing
 import os
 from pathlib import Path
-from typing import Dict, List
+from typing import List
 
 import numpy as np
 import torch
-from onnxruntime import GraphOptimizationLevel, InferenceSession, SessionOptions
-from onnxruntime.transformers import optimizer
-from onnxruntime.transformers.fusion_options import FusionOptions
-from onnxruntime.transformers.onnx_model_bert import BertOnnxModel
 from pycuda._driver import Stream
 from torch.cuda import get_device_name
 from torch.cuda.amp import autocast
 from transformers import (
     AutoModelForSequenceClassification,
-    PreTrainedModel,
+    PreTrainedModel, AutoTokenizer,
 )
 
-from tensorrt_utils import build_engine, save_engine, get_binding_idxs, infer_tensorrt
-from utils import print_timings, setup_logging, track_infer_time, prepare_input
+from backends.ort_utils import create_model_for_provider, convert_to_onnx, optimize_onnx
+from backends.trt_utils import build_engine, save_engine, get_binding_idxs, infer_tensorrt
+from benchmarks.utils import print_timings, setup_logging, track_infer_time, prepare_input
 from tensorrt.tensorrt import Logger, Runtime, IExecutionContext
 import tensorrt as trt
 import pycuda.driver as cuda
@@ -33,62 +29,15 @@ import pycuda.autoinit
 # TODO script shell to run all commands including the benchmark
 # TODO format code
 
-def create_model_for_provider(path: str, provider_to_use: str) -> InferenceSession:
-    options = SessionOptions()
-    options.graph_optimization_level = GraphOptimizationLevel.ORT_ENABLE_ALL
-    if type(provider_to_use) == list and provider_to_use == "CPUExecutionProvider":
-        options.intra_op_num_threads = multiprocessing.cpu_count()
-    if type(provider_to_use) != list:
-        provider_to_use = [provider_to_use]
-    return InferenceSession(path, options, providers=provider_to_use)
-
-
-def convert_to_onnx(model_pytorch: PreTrainedModel, output_path: str, inputs_pytorch: Dict[str, torch.Tensor]) -> None:
-    with torch.no_grad():
-        torch.onnx.export(
-            model_pytorch,  # model to optimize
-            args=(inputs_pytorch["input_ids"], inputs_pytorch["token_type_ids"], inputs_pytorch["attention_mask"]),  # tuple of multiple inputs
-            f=output_path,  # output path / file object
-            opset_version=12,  # the ONNX version to use
-            do_constant_folding=True,  # simplify model (replace constant expressions)
-            input_names=["input_ids", "token_type_ids", "attention_mask"],  # input names
-            output_names=["model_output"],  # output name
-            dynamic_axes={  # declare dynamix axis for each input / output (dynamic axis == variable length axis)
-                "input_ids": {0: "batch_size", 1: "sequence"},
-                "token_type_ids": {0: "batch_size", 1: "sequence"},
-                "attention_mask": {0: "batch_size", 1: "sequence"},
-                "model_output": {0: "batch_size"},
-            },
-            verbose=False,
-        )
-
-
-def optimize_onnx(onnx_path: str, onnx_optim_fp16_path: str, use_cuda: bool) -> None:
-    optimization_options = FusionOptions("bert")
-    optimization_options.enable_gelu_approximation = True  # additional optimization
-    optimized_model: BertOnnxModel = optimizer.optimize_model(
-        input=onnx_path,
-        model_type="bert",
-        use_gpu=use_cuda,
-        opt_level=1,
-        num_heads=0,  # automatic detection
-        hidden_size=0,  # automatic detection
-        optimization_options=optimization_options,
-    )
-
-    optimized_model.convert_float_to_float16()  # FP32 -> FP16
-    logging.info(f"optimizations applied: {optimized_model.get_fused_operator_statistics()}")
-    optimized_model.save_model_to_file(onnx_optim_fp16_path)
-
 
 setup_logging()
 
 
 def main():
     parser = argparse.ArgumentParser(description="optimize and deploy transformers", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument("-m", "--model", required=True, help="transofmer path or URL to Hugging Face Hub")
-    parser.add_argument("-b", "--batch-size", default=[1], help="batch size(s) to optimize for (min, opt, max)", type=int, nargs=3)
-    parser.add_argument("-s", "--seq-len", default=[16], help="sequence length(s) to optimize for (min, opt, max)", type=int, nargs=3)
+    parser.add_argument("-m", "--model", required=True, help="path to model or URL to Hugging Face Hub")
+    parser.add_argument("-b", "--batch-size", default=[1, 1, 1], help="batch sizes to optimize for (min, opt, max)", type=int, nargs=3)
+    parser.add_argument("-s", "--seq-len", default=[16, 16, 16], help="sequence lengths to optimize for (min, opt, max)", type=int, nargs=3)
     parser.add_argument("-w", "--workspace-size", default=10000, help="workspace size in MiB (TensorRT)", type=int)
     parser.add_argument("-o", "--output", default="triton_models", help="name to be used for ")
     parser.add_argument("-n", "--name", default="transformer", help="name to be used in triton server")
@@ -97,40 +46,25 @@ def main():
     parser.add_argument("--nb-measures", default=1000, help="# of inferences for benchmarks", type=int)
     args, _ = parser.parse_known_args()
 
-    nb_batch: List[int] = list()
-    seq_len: List[int] = list()
-    assert isinstance(args.batch_size, type(args.seq_len))
-    if isinstance(args.batch_size, int) and isinstance(args.seq_len, int):
-        nb_batch = [args.batch_size]
-        seq_len = [args.seq_len]
-    else:
-        nb_batch = args.batch_size
-        seq_len = args.seq_len
-
-    if len(nb_batch) == 1 and len(seq_len) == 1:
-        nb_batch *= 3
-        seq_len *= 3
-    # TODO improve
-    assert len(nb_batch) == 3 and len(seq_len) == 3, "provide 1 or 3 batch sizes / seq len"
-
     Path(args.output).mkdir(parents=True, exist_ok=True)
     onnx_model_path = os.path.join(args.output, "model-original.onnx")
-    # infered_shape_model_onnx_path = os.path.join(args.output, "model-shape.onnx")
     onnx_optim_fp16_path = os.path.join(args.output, "model.onnx")
+    # onnx_fp16_path = os.path.join(args.output, "model-half.onnx")
     tensorrt_path = os.path.join(args.output, "model.plan")
 
     assert torch.cuda.is_available(), "CUDA is not available. Please check your CUDA installation"
+    tokenizer = AutoTokenizer.from_pretrained(args.model)
+    input_names: List[str] = tokenizer.model_input_names
     model_pytorch: PreTrainedModel = AutoModelForSequenceClassification.from_pretrained(args.model)
     model_pytorch.cuda()
     model_pytorch.eval()
 
-    tensor_shapes = list(zip(nb_batch, seq_len))
-    inputs_pytorch, inputs_onnx = prepare_input(batch_size=tensor_shapes[-1][0], seq_len=tensor_shapes[-1][1], include_token_ids=True)
+    tensor_shapes = list(zip(args.batch_size, args.seq_len))
+    inputs_pytorch, inputs_onnx = prepare_input(batch_size=tensor_shapes[-1][0], seq_len=tensor_shapes[-1][1], include_token_ids="token_type_ids" in input_names)
 
-    with torch.no_grad():
-        output = model_pytorch(**inputs_pytorch)
-        output = output.logits  # extract the value of interest
-        output_pytorch: np.ndarray = output.detach().cpu().numpy()
+    output = model_pytorch(**inputs_pytorch)
+    output = output.logits  # extract the value of interest
+    output_pytorch: np.ndarray = output.detach().cpu().numpy()
 
     logging.info(f"[Pytorch] input shape {inputs_pytorch['input_ids'].shape}")
     logging.info(f"[Pytorch] output shape: {output_pytorch.shape}")
@@ -138,6 +72,7 @@ def main():
     convert_to_onnx(model_pytorch=model_pytorch, output_path=onnx_model_path, inputs_pytorch=inputs_pytorch)
     onnx_model = create_model_for_provider(path=onnx_model_path, provider_to_use="CUDAExecutionProvider")
     output_onnx = onnx_model.run(None, inputs_onnx)
+
     del onnx_model
 
     assert np.allclose(a=output_onnx, b=output_pytorch, atol=1e-1)
@@ -158,7 +93,7 @@ def main():
 
     trt_logger: Logger = trt.Logger(trt.Logger.INFO)
     runtime: Runtime = trt.Runtime(trt_logger)
-    engine = build_engine(runtime=runtime, onnx_file_path=onnx_model_path, logger=trt_logger, min_shape=tensor_shapes[0], optimal_shape=tensor_shapes[1], max_shape=tensor_shapes[2], workspace_size=args.workspace_size * 1024)
+    engine = build_engine(runtime=runtime, onnx_file_path=onnx_model_path, logger=trt_logger, min_shape=tensor_shapes[0], optimal_shape=tensor_shapes[1], max_shape=tensor_shapes[2], workspace_size=args.workspace_size * 1024 * 1024)
     save_engine(engine=engine, engine_file_path=tensorrt_path)
     # noinspection DuplicatedCode
     stream: Stream = pycuda.driver.Stream()
@@ -168,7 +103,10 @@ def main():
     # retrieve input/output IDs
     input_binding_idxs, output_binding_idxs = get_binding_idxs(engine, profile_index)  # type: List[int], List[int]
     tensorrt_output = infer_tensorrt(context=context, host_inputs=inputs_onnx, input_binding_idxs=input_binding_idxs, output_binding_idxs=output_binding_idxs, stream=stream)
-    assert np.allclose(a=tensorrt_output, b=output_pytorch, atol=1e-1)
+    assert np.allclose(a=tensorrt_output, b=output_pytorch, atol=1e-1), f"tensorrt accuracy is too low:\n" \
+                                                                        f"Pythorch:\n{output_pytorch}\n" \
+                                                                        f"VS\n" \
+                                                                        f"TensorRT:\n{tensorrt_output}"
 
     timings = {}
     for _ in range(args.warmup):
@@ -197,22 +135,23 @@ def main():
     del model
 
     if args.pytorch:
-        for _ in range(args.warmup):
-            _ = model_pytorch(**inputs_pytorch)
-        time_buffer = []
-        for _ in range(args.nb_measures):
-            with track_infer_time(time_buffer):
-                model_pytorch(**inputs_pytorch)
-        timings["Pytorch_fp32"] = time_buffer
-
-        with autocast():
+        with torch.inference_mode():
             for _ in range(args.warmup):
-                model_pytorch(**inputs_pytorch)
+                _ = model_pytorch(**inputs_pytorch)
             time_buffer = []
             for _ in range(args.nb_measures):
                 with track_infer_time(time_buffer):
                     model_pytorch(**inputs_pytorch)
-            timings["Pytorch_fp16"] = time_buffer
+            timings["Pytorch_fp32"] = time_buffer
+
+            with autocast():
+                for _ in range(args.warmup):
+                    model_pytorch(**inputs_pytorch)
+                time_buffer = []
+                for _ in range(args.nb_measures):
+                    with track_infer_time(time_buffer):
+                        model_pytorch(**inputs_pytorch)
+                timings["Pytorch_fp16"] = time_buffer
 
     logging.info(f"inference done on {get_device_name(0)}")
