+++ requirements.txt
@@ -7,7 +7,7 @@ transformers
 sentencepiece
 tritonclient[all]
 onnx
-onnxruntime-gpu==1.9.0
+onnxruntime-gpu
 sympy
 coloredlogs
 pycuda
